get_injured moved to conditions.py
