+++ scripts/events_module/pregnancy/handle_already_pregnant.py
@@ -9,6 +9,7 @@ from scripts.cat.names import Name
 from scripts.cat_relations.relationship import Relationship
 from scripts.clan_package.get_clan_cats import find_alive_cats_with_rank
 from scripts.clan_package.settings import get_clan_setting
+from scripts.conditions import get_injured
 from scripts.config import get_config
 from scripts.event_class import Single_Event
 from scripts.events_module.consequences import (
@@ -247,7 +248,7 @@ def handle_two_moon_pregnant(cat: Cat):
             death_event = i18n.t("conditions.pregnancy.kitting_death", name=cat.name)
         cat.history.add_death(death_text=death_event)
     elif not cat.status.is_outsider:  # if cat doesn't die, give recovering from birth
-        cat.get_injured("recovering from birth", event_triggered=True)
+        get_injured(cat, "recovering from birth", event_triggered=True)
         if "blood loss" in cat.injuries:
             if cat.status.is_leader:
                 death_event = i18n.t("conditions.pregnancy.leader_kitting_death_severe")
