+++ scripts/events_module/text_pool_event/handle_consequences.py
@@ -17,6 +17,7 @@ from scripts.clan_resources.freshkill import (
     HUNTER_BONUS,
     HUNTER_EXP_BONUS,
 )
+from scripts.conditions import get_ill, get_injured
 from scripts.config import get_config
 from scripts.events_module.consequences import unpack_rel_block, check_stolen_vitality
 from scripts.events_module.future.prep_and_trigger import prep_future_event
@@ -351,7 +352,7 @@ def _handle_conditions(
             chosen_condition = choice(list(conditions_for_cat))
 
             if chosen_condition in INJURIES:
-                c.get_injured(chosen_condition, lethal=lethal, potential_scars=scars)
+                get_injured(c, chosen_condition, lethal=lethal, potential_scars=scars)
             elif chosen_condition in ILLNESSES:
                 get_ill(c, chosen_condition, lethal=lethal)
             else:
